requirements.txt
wheel
Discord.py
requests